+++ js/base.js
@@ -1,0 +1,58 @@
+const HEADERS = [
+  'Шикарный дом на берегу залива',
+  'Уютная квартира для молодой пары',
+  'Бунгало для двоих',
+  'Дворец в старо-японском стиле',
+  'Дом для большой семьи',
+  'Квартира в тихом районе',
+  'Квартира с видом на парк Ёёги',
+  'Восточный дворец',
+  'Дом с террасой',
+  'Бунгало. Эконом-вариант.',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+// Check-in & Check-out are the same
+const CHECKTIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTIONS = [
+  'Здесь есть все самое необходимое.',
+  'Лучшее предложение в этом районе.',
+  'Можно с животными.',
+  'Вам здесь понравится!',
+  'Отличный вид из панорамных окон',
+  'Имеется большая кладовка',
+  'Гараж на две машины',
+  'В пешей доступности от метро',
+  'Развитая инфраструктура',
+  'Окна выходят на запад. Вы сможете наслаждаться закатами каждый вечер!',
+];
+
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const SIMILAR_AD_AMOUNT = 10;
+
+export {HEADERS, TYPES, CHECKTIME, FEATURES, DESCRIPTIONS, PHOTOS, SIMILAR_AD_AMOUNT}
